Minore fix

+++ Sources/TraderUserDto/Request/UserPatchDto.swift
@@ -12,14 +12,14 @@ public struct UserPatchDto: Codable, Sendable {
     public let isBlock: Bool?
     public let firstName: String?
     public let lastName: String?
-    public let roleLevel: Int?
+    public let roleLevel: UInt8?
 
     // MARK: - Init
     public init(
         isBlock: Bool? = nil,
         firstName: String? = nil,
         lastName: String? = nil,
-		roleLevel: Int? = nil
+		roleLevel: UInt8? = nil
     ) {
         self.isBlock = isBlock
         self.firstName = firstName
@@ -51,7 +51,7 @@ extension UserPatchDto {
 			isBlock: values[0] == "" ? nil : values[0] == "1",
 			firstName: values[1] == "" ? nil : values[1],
 			lastName: values[2] == "" ? nil : values[2],
-			roleLevel: Int(values[3])
+			roleLevel: UInt8(values[3])
 		)
 	}
 }
